requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
comm==0.2.2
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.1.6
django-cors-headers==4.6.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
executing==2.2.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
msgpack==1.1.0
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
psycopg==3.2.4
psycopg-binary==3.2.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pyzmq==26.2.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
zope.interface==7.2
